Added Read Bin and fixed dependency of 3ds
added functionality to read bin files from java code, got rid of dependency of nanonispy27
